gh-3056 Move HexDump to util-shared, fix checkstyle
