+++ stroom-util-shared/src/main/java/stroom/util/shared/string/HexDump.java
@@ -1,0 +1,119 @@
+package stroom.util.shared.string;
+
+import stroom.util.shared.Range;
+
+import java.io.InputStream;
+import java.nio.charset.Charset;
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+
+public class HexDump {
+
+    public static final int MAX_BYTES_PER_LINE = 32;
+    // This depends on how we render the line of MAX_BYTES_PER_LINE (includes the  at the end \n)
+    public static final int MAX_CHARS_PER_DUMP_LINE = 149;
+
+    private final List<HexDumpLine> lines;
+    // The charset of the decoded bytes.
+    private final Charset charset;
+    // The range covered by all lines in this hex dump
+    private final Range<Long> byteOffsetRange;
+
+    public HexDump(final List<HexDumpLine> lines,
+                   final Charset charset,
+                   final Range<Long> byteOffsetRange) {
+        this.lines = lines;
+        this.charset = charset;
+        this.byteOffsetRange = byteOffsetRange;
+    }
+
+    public List<HexDumpLine> getLines() {
+        return lines;
+    }
+
+    public boolean isEmpty() {
+        return lines.isEmpty();
+    }
+
+    public Charset getCharset() {
+        return charset;
+    }
+
+    public Optional<HexDumpLine> getLine(final int lineNo) {
+        if (lines.isEmpty()) {
+            return Optional.empty();
+        } else {
+            final long firstLineNo = lines.get(0).getLineNo();
+            if (lineNo < firstLineNo) {
+                return Optional.empty();
+            } else {
+                final long idx = lineNo - firstLineNo;
+                if (idx >= lines.size()) {
+                    return Optional.empty();
+                } else {
+                    return Optional.ofNullable(lines.get((int) idx));
+                }
+            }
+        }
+    }
+
+    public int getLineCount() {
+        return lines.size();
+    }
+
+    public Optional<HexDumpLine> getFirstLine() {
+        if (lines.isEmpty()) {
+            return Optional.empty();
+        } else {
+            return Optional.of(lines.get(0));
+        }
+    }
+
+    public Optional<HexDumpLine> getLastLine() {
+        if (lines.isEmpty()) {
+            return Optional.empty();
+        } else {
+            return Optional.of(lines.get(lines.size() - 1));
+        }
+    }
+
+    /**
+     * @return Number of bytes rendered by the hex dump.
+     */
+    public long getDumpByteCount() {
+        return byteOffsetRange.size().longValue();
+    }
+
+    /**
+     * @return The range of bytes rendered by the hex dump, e.g. if only a subset of
+     * an {@link InputStream} is dumped.
+     */
+    public Range<Long> getByteOffsetRange() {
+        return byteOffsetRange;
+    }
+
+    /**
+     * @return The size of the hex dump when output as a string. Not the
+     * number of chars rendered by the hex dump.
+     */
+    public long getDumpCharCount() {
+        return lines.stream()
+                .mapToInt(HexDumpLine::getDumpLineCharCount)
+                .sum();
+    }
+
+    public String getHexDumpAsStr() {
+        return lines.stream()
+                .map(HexDumpLine::getLine)
+                .collect(Collectors.joining("\n"));
+    }
+
+    @Override
+    public String toString() {
+        return "HexDump{" +
+                "lines=" + lines +
+                ", charset=" + charset +
+                '}';
+    }
+}
